Separate load puzzle cards from metric cards
