+++ src/app/load-puzzle/load-puzzle.component.scss
@@ -10,8 +10,7 @@
     }
   }
 
-  .card.existing {
-    width: 200px;
-    height: 200px;
+  .text-truncate {
+    max-width: 100%;
   }
 }
